+++ packages/patrol_cli/test/src/mocks.dart
@@ -1,5 +1,6 @@
 import 'dart:io' as io;
 
+import 'package:adb/adb.dart';
 import 'package:http/http.dart' as http;
 import 'package:mocktail/mocktail.dart';
 import 'package:patrol_cli/src/analytics/analytics.dart';
@@ -9,6 +10,7 @@ import 'package:patrol_cli/src/compatibility_checker/compatibility_checker.dart'
 import 'package:patrol_cli/src/dart_defines_reader.dart';
 import 'package:patrol_cli/src/devices.dart';
 import 'package:patrol_cli/src/ios/ios_test_backend.dart';
+import 'package:patrol_cli/src/macos/macos_test_backend.dart';
 import 'package:patrol_cli/src/pubspec_reader.dart';
 import 'package:patrol_cli/src/test_bundler.dart';
 import 'package:patrol_cli/src/test_finder.dart';
@@ -35,6 +37,8 @@ class MockAndroidTestBackend extends Mock implements AndroidTestBackend {}
 
 class MockIOSTestBackend extends Mock implements IOSTestBackend {}
 
+class MockMacOSTestBackend extends Mock implements MacOSTestBackend {}
+
 class MockTestFinder extends Mock implements TestFinder {}
 
 class MockTestBundler extends Mock implements TestBundler {}
@@ -44,6 +48,8 @@ class MockDartDefinesReader extends Mock implements DartDefinesReader {}
 class MockPubspecReader extends Mock implements PubspecReader {}
 
 class MockCompatibilityChecker extends Mock implements CompatibilityChecker {}
+
+class MockAdb extends Mock implements Adb {}
 
 class MockAnalytics extends Mock implements Analytics {
   @override